add readconfirm

+++ fish/functions/cleanDocker.fish
@@ -1,6 +1,8 @@
 function cleanDocker
-    if docker ps -a -q | count > 0
-        docker stop (docker ps -a -q)
-        docker rm (docker ps -a -q)
+    if read_confirm "Do you want to stop and remove all containers?"
+        if docker ps -a -q | count > 0
+            docker stop (docker ps -a -q)
+            docker rm (docker ps -a -q)
+        end
     end
 end
